fix: return 0 when command is not run.
We were indicating whether the command was ran by returning different
exist codes for testing. However, this makes tools like just think the
command failed even when it runs as expected. We will have to
communicate this information some other way, probably by adding an
option to print the information.