feat: dependencies, targets, and a new "sentinal" files can be given as options
added --dependency and --target options to allow specifying deps and
dependencie with options instead of arguments. Also added a new
"sentinal" type. If a sentinal file/directory is given, the command will
be run if the sentinal exists. Sort of like the oposite of a target.

hmmm... would the opposite of a depdency, run only if it has not
changed, make sense...