feat: added --shell option to allow running command with shell=True
If you try to run a command that uses a shell builtin you will get an
error running it with Popen unless you pass shell=True. The new command
line option allows you to pass this in for those cases it is needed.

Ran into this on Windows when trying to run 'echo' in powershell,
apparently its a builtin.