feat: change detection faster with blake3 hash
Switched to using blake3 over openssl md5. When using the md_5
hash, we were significantly slower than the old Python version
of this tool. Switching to openssl's md5 made us the same speed.
I guess openssl's code is highly optimized? I think it's written
in C/Assymbly, but that means you have to have the openssl
dev package installed to build run-if

Computing hashes for files in parallel made us faster than the Python
version.

Blake3 is faster than md5 and the Rust implementation is faster than
openssl's md5, even for debug builds. So we switched hash functions.
Even if it was the _same_ speed, switching to Blake3 would be an
improvement because is implemented in pure rust so we don't
have to install external dependencies to build.